+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/CombinationGraph/BasicCombinationGraphs/BarGraphWithLine.jsx
@@ -6,13 +6,17 @@ import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Bar/BarGraph.module.scss';
 import ExampleGraphContainer from '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer';
 import getCombinationGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Combination/combinationDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/BarGraphWithLineData';
+import lineDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/lineDataset1';
+import barDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/barDataset';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
 const graphConfig = utils.deepClone(getCombinationGraphConfig('#barGraphWithLine'));
-const dataset = utils.deepClone(data);
+const dataset = [
+  utils.deepClone(barDataset),
+  utils.deepClone(lineDataset),
+];
 
 const BarGraphWithLineCombinationExample = () => {
   React.useEffect(() => {