+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/CombinationGraph/BasicCombinationGraphs/LineWithPairedResult.jsx
@@ -1,18 +1,22 @@
 import React from 'react';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Graph.module.scss';
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph.module.scss';
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/PairedResult/PairedResultGraph.module.scss';
 import ExampleGraphContainer from '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer';
 import lineAndPairedResultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Combination/lineAndPairedResultConfig';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/lineAndPairedResultData';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import lineDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/lineDataset2';
+import pairedResultDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/pairedResultDataset';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
 const graphConfig = utils.deepClone(lineAndPairedResultConfig('#lineWithPairedResult'));
-const dataset = utils.deepClone(data);
+const dataset = [
+  utils.deepClone(lineDataset),
+  utils.deepClone(pairedResultDataset),
+];
 
 const LineWithPairedResultCombinationExample = () => {
   React.useEffect(() => {