+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/CombinationGraph/BasicCombinationGraphs/BarGraphWithLine.jsx
@@ -3,13 +3,17 @@ import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import CombinationGraph from '@cerner/terra-graphs/lib/components/Combination/CombinationGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
 import getCombinationGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Combination/combinationDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/BarGraphWithLineData';
+import lineDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/lineDataset1';
+import barDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/barDataset';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
 const graphConfig = utils.deepClone(getCombinationGraphConfig('#barGraphWithLine'));
-const dataset = utils.deepClone(data);
+const dataset = [
+  utils.deepClone(barDataset),
+  utils.deepClone(lineDataset),
+];
 
 export default () => (
   <>