+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/CombinationGraph/BasicCombinationGraphs/LineWithPairedResult.jsx
@@ -3,13 +3,17 @@ import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import CombinationGraph from '@cerner/terra-graphs/lib/components/Combination/CombinationGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
 import lineAndPairedResultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Combination/lineAndPairedResultConfig';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/lineAndPairedResultData';
+import lineDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/lineDataset2';
+import pairedResultDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/pairedResultDataset';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
 const graphConfig = utils.deepClone(lineAndPairedResultConfig('#lineWithPairedResult'));
-const dataset = utils.deepClone(data);
+const dataset = [
+  utils.deepClone(lineDataset),
+  utils.deepClone(pairedResultDataset),
+];
 
 export default () => (
   <>