+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/Combination/Regions/Simple.test.jsx
@@ -3,35 +3,14 @@ import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import CombinationGraph from '@cerner/terra-graphs/lib/components/Combination/CombinationGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
 import lineAndPairedResultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Combination/lineAndPairedResultConfig';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/lineAndPairedResultData';
+import lineDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/lineSimpleRegionDataset';
+import pairedResultDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/pairedResultSimpleRegionDataset';
 
 const graphConfig = utils.deepClone(lineAndPairedResultConfig('#lineWithPairedResultRegion'));
-const dataset = utils.deepClone(data);
-
-dataset[0].regions = [
-  {
-    axis: 'y',
-    start: 6,
-    end: 18,
-  },
+const dataset = [
+  utils.deepClone(lineDataset),
+  utils.deepClone(pairedResultDataset),
 ];
-dataset[1].regions = {
-  high: [
-    {
-      axis: 'y2',
-      start: 120,
-      end: 180,
-      color: '#c8cacb',
-    },
-  ],
-  low: [
-    {
-      axis: 'y2',
-      start: 20,
-      end: 80,
-    },
-  ],
-};
 
 export default () => (
   <>