+++ packages/terra-graphs/CHANGELOG.md
@@ -28,5 +28,6 @@
   * Updated Gantt Chart snapshots.
   * Updated Timeline graph snapshots.
   * Updated all snapshots.
+  * Updated Combination graph snapshots.
 
 * Initial stable release